removed build-data.js and moved writeStreams to appropriate modules
